A add a file system management tool for ZenFS
Summary:

This adds a file system management tool for ZenFS. It is required
for setting up the metadata for the filesystem.

It also allows the user to list files on the file system and
can be extended in the future to do other management tasks
like offline garbage collection, backups etc.

Examples:

1. Create a zenfs file system on zoned block device /dev/nullb1 with
auxiliary file storage (logfiles, lockfiles) under /tmp/zenfs_nullb1 and
with a finishing threshold for zones of 5%. If the zone has less than
finish_threshold capacity left, no additional extents will be mapped
to the zone.

./zenfs mkfs --zbd=nullb1 --aux_path=/tmp/zenfs_nullb1 --finish_threshold=5
ZenFS file system created. Free space: 31744 MB

2. List files:

./zenfs list --zbd=nullb1 --path=/rocksdbtest/dbbench
/rocksdbtest/dbbench/LOG
/rocksdbtest/dbbench/LOCK
/rocksdbtest/dbbench/000003.log
/rocksdbtest/dbbench/CURRENT
/rocksdbtest/dbbench/IDENTITY
/rocksdbtest/dbbench/MANIFEST-000001
/rocksdbtest/dbbench/OPTIONS-000005